requirements.txt
argparse==1.2.1
caniusepython3==2.1.0
distlib==0.1.8
dj-database-url==0.3.0
dj-email-url==0.0.2
django-admin-honeypot==0.3.0
django-cache-url==0.8.0
django-configurations==0.8
django-easy-pjax==1.1.0
django-pgfields==1.4.3
django-picklefield==0.3.1
django-pipeline==1.3.24
django-rq==0.6.1
django-redis==3.6.2
django-secure==1.0
django-templates-macros==0.1
django-vanilla-views==1.0.3
Django==1.6.4
djangorestframework==2.3.13
futures==2.1.6
hiredis==0.1.3
HireFire==0.2
mock==1.0.1
pathlib==1.0
pip-tools==0.3.4
psycopg2==2.5.2
python-dateutil==2.2
pytz==2014.2
raven==4.2.1
redis==2.9.1
requests==2.2.1
rq==0.4.5
six==1.6.1
South==0.8.4
times==0.6.2
waitress==0.8.8
whitenoise==1.0.2
https://github.com/jezdez/django-templates-macros/archive/master.zip
